git-svn-id: file:///usr/local/svn/fsc2@8224 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/src/fsc2.h
@@ -377,7 +377,7 @@ struct Crash_Info {
     volatile sig_atomic_t signo;  /* signal indicating the type of crash */
     void *address;                /* address the crash happened */
     void *trace[ MAX_TRACE_LEN ]; /* addresses of backtrace */
-    int trace_length;             /* length of backtrace */
+    size_t trace_length;          /* length of backtrace */
 };
